Refactor some methods to be under InfoHandler
