+++ src/main/java/seedu/address/logic/commands/GenerateLocationsCommand.java
@@ -1,6 +1,7 @@
 package seedu.address.logic.commands;
 
 import static java.util.Objects.requireNonNull;
+import static seedu.address.commons.core.Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX;
 
 import java.util.List;
 
@@ -8,6 +9,7 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.ModelPredicate;
+import seedu.address.model.VisitBook;
 
 /**
  * Finds and lists all persons in address book whose name contains any of the argument keywords.
@@ -22,6 +24,9 @@ public class GenerateLocationsCommand extends Command {
             + "Parameters: PERSONID\n"
             + "Example: " + COMMAND_WORD + " 1";
 
+    public static final String MESSAGE_PERSON_HAS_NO_VISITS = "This person is not associated with any visits";
+    public static final String MESSAGE_PERSON_IS_NOT_INFECTED = "This person is not infected";
+
     private final Index personId;
 
     public GenerateLocationsCommand(Index personId) {
@@ -31,7 +36,18 @@ public class GenerateLocationsCommand extends Command {
     @Override
     public CommandResult execute(Model model) throws CommandException {
         requireNonNull(model);
-        List<Integer> locationIds = CommandUtil.generateLocationIdsByPerson(model, personId);
+        if (personId.getZeroBased() >= model.getAddressBook().getPersonList().size()) {
+            throw new CommandException(MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
+        }
+        if (!model.getAddressBook().getPersonList()
+                .get(personId.getZeroBased()).getInfectionStatus().getStatusAsBoolean()) {
+            throw new CommandException(MESSAGE_PERSON_IS_NOT_INFECTED);
+        }
+        VisitBook visitsByPerson = model.getInfoHandler().generateVisitsByPerson(personId);
+        if (visitsByPerson.getVisitList().isEmpty()) {
+            throw new CommandException(MESSAGE_PERSON_HAS_NO_VISITS);
+        }
+        List<Integer> locationIds = model.getInfoHandler().generateLocationIdsByVisitBook(visitsByPerson);
         model.updateFilteredLocationList(ModelPredicate.getPredicateShowLocationsById(locationIds));
         return new CommandResult(
                 "Generated locations for: " + model.getAddressBook()
