+++ src/main/java/seedu/address/logic/commands/GeneratePeopleCommand.java
@@ -1,17 +1,15 @@
 package seedu.address.logic.commands;
 
 import static java.util.Objects.requireNonNull;
+import static seedu.address.commons.core.Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.ModelPredicate;
-import seedu.address.model.ReadOnlyVisitBook;
 import seedu.address.model.VisitBook;
-import seedu.address.model.visit.Visit;
 
 /**
  * Finds and lists all persons in address book whose name contains any of the argument keywords.
@@ -27,6 +25,8 @@ public class GeneratePeopleCommand extends Command {
             + "Example: " + COMMAND_WORD + " 1";
 
     public static final String MESSAGE_NO_PEOPLE_FOUND = "There were no people in contact with the given person";
+    public static final String MESSAGE_PERSON_HAS_NO_VISITS = "This person is not associated with any visits";
+    public static final String MESSAGE_PERSON_IS_NOT_INFECTED = "This person is not infected";
 
     private final Index personId;
 
@@ -37,25 +37,20 @@ public class GeneratePeopleCommand extends Command {
     @Override
     public CommandResult execute(Model model) throws CommandException {
         requireNonNull(model);
-        List<Integer> locationIds = CommandUtil.generateLocationIdsByPerson(model, personId);
-        List<Integer> personIds = new ArrayList<>();
-        ReadOnlyVisitBook tempVisitBook = model.getVisitBook();
-        VisitBook affectedVisits = new VisitBook();
-        for (Integer locationId : locationIds) {
-            Index locationIndex = Index.fromZeroBased(locationId);
-            for (int i = 0; i < tempVisitBook.getVisitList().size(); i++) {
-                Visit visit = tempVisitBook.getVisitList().get(i);
-                if (visit.getLocationId().equals(locationIndex)) {
-                    affectedVisits.addVisit(visit);
-                }
-            }
+        if (personId.getZeroBased() >= model.getAddressBook().getPersonList().size()) {
+            throw new CommandException(MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
         }
-        for (Visit visit : affectedVisits.getVisitList()) {
-            if (visit.getPersonId().getZeroBased() == personId.getZeroBased()) {
-                continue;
-            }
-            personIds.add(visit.getPersonId().getZeroBased());
+        if (!model.getAddressBook().getPersonList()
+                .get(personId.getZeroBased()).getInfectionStatus().getStatusAsBoolean()) {
+            throw new CommandException(MESSAGE_PERSON_IS_NOT_INFECTED);
         }
+        VisitBook visitsByPerson = model.getInfoHandler().generateVisitsByPerson(personId);
+        if (visitsByPerson.getVisitList().isEmpty()) {
+            throw new CommandException(MESSAGE_PERSON_HAS_NO_VISITS);
+        }
+        List<Integer> locationIds = model.getInfoHandler().generateLocationIdsByVisitBook(visitsByPerson);
+        VisitBook affectedVisits = model.getInfoHandler().generateVisitsByLocationIds(locationIds);
+        List<Integer> personIds = model.getInfoHandler().generatePersonIdsByVisitBook(affectedVisits, personId);
         if (personIds.isEmpty()) {
             throw new CommandException(MESSAGE_NO_PEOPLE_FOUND);
         }
